--- a/index.js
+++ b/index.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const URL = require('url').URL
+const PROTOCOL_PLACEHOLDER = 'protocol-placeholder:'
+const { URL, parse } = require('url')
 const parseForwarded = require('forwarded-parse')
 const net = require('net')
 
@@ -9,10 +10,10 @@ module.exports = function (req) {
   const url = parsePartialURL(raw)
   const secure = req.secure || (req.connection && req.connection.encrypted)
   const result = { raw: raw }
-  let host
+  let host, forwarded
 
   if (req.headers.forwarded) {
-    let forwarded = getFirstHeader(req, 'forwarded')
+    forwarded = getFirstHeader(req, 'forwarded')
     try {
       // Always choose the original (first) Forwarded pair in case the request
       // passed through multiple proxies
@@ -23,7 +24,6 @@ module.exports = function (req) {
         const port = conn[conn.length - 1].split(':')[1]
         if (port) host.port = Number(port)
       }
-      if (forwarded.proto) host.protocol = forwarded.proto + ':'
     } catch (e) {}
   } else if (req.headers['x-forwarded-host']) {
     host = parsePartialURL(getFirstHeader(req, 'x-forwarded-host'))
@@ -38,13 +38,14 @@ module.exports = function (req) {
   }
 
   // protocol
-  if (url.protocol) result.protocol = url.protocol
+  if (url.protocol !== PROTOCOL_PLACEHOLDER) result.protocol = url.protocol
   else if (req.headers['x-forwarded-proto']) result.protocol = getFirstHeader(req, 'x-forwarded-proto') + ':'
   else if (req.headers['x-forwarded-protocol']) result.protocol = getFirstHeader(req, 'x-forwarded-protocol') + ':'
   else if (req.headers['x-url-scheme']) result.protocol = getFirstHeader(req, 'x-url-scheme') + ':'
   else if (req.headers['front-end-https']) result.protocol = getFirstHeader(req, 'front-end-https') === 'on' ? 'https:' : 'http:'
   else if (req.headers['x-forwarded-ssl']) result.protocol = getFirstHeader(req, 'x-forwarded-ssl') === 'on' ? 'https:' : 'http:'
-  else if (host.protocol) result.protocol = host.protocol
+  else if (forwarded && forwarded.proto) result.protocol = forwarded.proto + ':'
+  else if (host.protocol && host.protocol !== PROTOCOL_PLACEHOLDER) result.protocol = host.protocol
   else if (secure) result.protocol = 'https:'
   else result.protocol = 'http:'
 
@@ -92,14 +93,10 @@ function getFirstHeader (req, header) {
 
 function parsePartialURL (url) {
   const containsProtocol = url.indexOf('://') !== -1
-  const urlInstance = new URL(containsProtocol ? url : 'invalid://' + url)
-  const result = {}
-  for (const prop in urlInstance) {
-    if (typeof urlInstance[prop] !== 'function') result[prop] = urlInstance[prop]
+  const urlToParse = containsProtocol ? url : PROTOCOL_PLACEHOLDER + '//' + url
+  try {
+    return new URL(urlToParse)
+  } catch (e) {
+    return parse(urlToParse)
   }
-  if (!containsProtocol) {
-    result.protocol = ''
-    result.href = result.href.replace('invalid://', '')
-  }
-  return result
 }