Replace the legacy `parse` API from `url` module and use the `URL` constructor instead. Reasons for this change are:

- API is deprecated and it use discouraged in [the documentation](https://nodejs.org/dist/latest-v18.x/docs/api/url.html#urlparseurlstring-parsequerystring-slashesdenotehost)
- Consumers of the package have discovered issues related to the parsing algorithm
    - URLs starting with double slash `//some/path` are not parsed correctly
    - URLs with auth like form `//user@foo` have the same problem

The solution proposed in this PR considers the raw URL as partial like the ones seek on the headers and adding a fake protocol if necessary (to remove later). The function iterates over the URL instance properties to fill a result object since some properties of URL do not accept empty values like protocol.

```javascript
const url = new URL('https://localhost:8080/path')
console.log(url.protocol) // prints "https:"
url.protocol = '';
console.log(url.protocol) // keeps printing "https:"
```

Tests have been added in  http/https requests to test these specific scenarios.