@styfle 2023-03-02
Which version of Node.js did you benchmark? 

There were some recent performance improvements that landed 3 weeks ago:

- https://github.com/nodejs/performance/issues/33