@trentm 2023-03-02
[I'm pulling this out of [the review comment](https://github.com/watson/original-url/pull/11#pullrequestreview-1302371395) so the conversation doesn't get broken up and more easily lost.]

> Do you mean to use a value which is longer and more unlikely to be used by consumers? Something like original-url-protocol: or something of the sorts?

Yes, that is what I had meant. It isn't very satisfying, however. :)

In [the nodejs-land issue](https://github.com/nodejs/web-server-frameworks/issues/71) that @styfle referred to, there was mention of a separate https://www.npmjs.com/package/request-target package for handling parsing a `req.url`. Perhaps it would be preferable to use that (or logic similar to it, but *not* `url.parse()`) for parsing the input `req.url`?  I haven't dug into it.

The parsing of the other headers (`Host`, `Forwarded`, et al) could still perhaps use `new URL` with a try-catch guard, assuming those headers are specified to have, or typically have *complete URLs*.

> ... IMO this is a breaking change and needs to be reflected in the version with a major bump.

Yes, agreed. I'm not opposed to a new major rev for this package. It would then be nice to then get specific about what fields of the returned object are promised -- as opposed to the currently looser "The result object will also contain any other property normally returned by the Node.js core [url.parse](https://nodejs.org/api/url.html#url_url_parse_urlstring_parsequerystring_slashesdenotehost) function."