@david-luna 2023-03-13
I've checked  https://www.npmjs.com/package/request-target package and it's using a set of `RegExp` to match the string extracted from `req.url`. The expressions are derived from the syntax specified in [RFC3986](https://www.rfc-editor.org/rfc/rfc3986#appendix-A). I think we can add it as a dependency.

`WATHWG URL` and `request-target` are stricter than `url.parse()`. When the URL string is invalid the former throws and the later returns null. I think we should unify this but not sure which option is best. With this set of changes we will do breaking changes anyway but I wonder what would be better for package consumers.