@david-luna 2023-03-21
Using a non standard protocol in `new URL()` constructor changes the parsing result to a better detection of paths (starting with `/` and `//`) increasing the number of URLs being parsed without throwing. New results


Parse results for **/**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | http: |  | / |
| new-url | protocol: |  | / |
| url-parse |  |  | / |

Parse results for **/some/path**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | http: |  | /some/path |
| new-url | protocol: |  | /some/path |
| url-parse |  |  | /some/path |

Parse results for **//some/path**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | N/A | N/A | N/A |
| new-url | protocol: |  | //some/path |
| url-parse |  |  | //some/path |

Parse results for **/user@pass**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | http: |  | /user@pass |
| new-url | protocol: |  | /user@pass |
| url-parse |  |  | /user@pass |

Parse results for **//user@pass**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | N/A | N/A | N/A |
| new-url | protocol: |  | //user@pass |
| url-parse |  |  | //user@pass |

Parse results for **/some/path?key=value**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | http: |  | /some/path |
| new-url | protocol: |  | /some/path |
| url-parse |  |  | /some/path |

Parse results for **@@**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | N/A | N/A | N/A |
| new-url | N/A | N/A | N/A |
| url-parse |  |  | @@ |

Parse results for **///**
| method | protocol | host | pathname |
| - | - | - | - |
| request-target | N/A | N/A | N/A |
| new-url | protocol: |  | /// |
| url-parse |  |  | /// |

Urls parsed by method
| request-target | new-url | url-parse | total |
| - | - | - | - |
| 4 | 7 | 8 | 8 |