@david-luna 2023-03-21
The batches show that there are edge cases where `url.parse` still gives a result but we can cover most of them with `request-target` and `new URL(...)` being the former faster than the later. IMHO a good approach would be
- use `request-target`
- fallback to `new URL(...)`
- return null if both failed to parse